feat: only bind module prototype if it exists
Typescript 4.2 fixed the return type for Reflect.getPrototypeOf to
include null, which requires a check before binding the prototypical
methods on a module. This should not be a breaking change, and may
fix a runtime TypeError when binding modules without a prototype.
Technically this is a fix, but due to the potential impact, is being
marked as a minor release.